requirements.txt
pybirdbuddy==0.0.11
slack_sdk==3.19.5
python-dotenv==0.21.1